+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## Unreleased
+
+### Fixed
+
+- Notifications opening Alarms instead of Current State (#139, thanks @matiamb).
+
+### Improved
+
+- More reliable voltage reading (#142).
+
 ## 1.7.1
 
 1.7.0 was the last feature-oriented release before version 2.0 (October/November). The next releases will be less frequent and will consist only of bug fixes and small improvements.